Sub-task 2: Service interaction

+++ post/src/main/java/com/services/post/service/PostService.java
@@ -1,5 +1,6 @@
 package com.services.post.service;
 
+import com.services.post.client.UserServiceClient;
 import com.services.post.domain.Post;
 import com.services.post.repository.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +15,12 @@ public class PostService {
 
     private final PostRepository postRepository;
 
+    private final UserServiceClient userServiceClient;
+
     @Autowired
-    public PostService(PostRepository postRepository) {
+    public PostService(PostRepository postRepository, UserServiceClient userServiceClient) {
         this.postRepository = postRepository;
+        this.userServiceClient = userServiceClient;
     }
 
     @Transactional(readOnly = true)
@@ -31,17 +35,28 @@ public class PostService {
 
     @Transactional
     public Post create(Post post) {
-        return postRepository.save(post);
+        post = postRepository.save(post);
+        updateNumberOfPosts(post.getAuthorId(), 1);
+        return post;
     }
 
     @Transactional
     public void deleteById(long id) {
-        postRepository.deleteById(id);
+        Optional<Post> postOptional = postRepository.findById(id);
+        if (postOptional.isPresent()) {
+            Post post = postOptional.get();
+            postRepository.deleteById(id);
+            updateNumberOfPosts(post.getAuthorId(), -1);
+        }
     }
 
     @Transactional
     public Optional<Post> updateTextAndDateById(long id, String text) {
         postRepository.updateTextAndDateById(id, text, LocalDateTime.now());
         return postRepository.findById(id);
+    }
+
+    private void updateNumberOfPosts(long authorId, int amountChange) {
+        userServiceClient.updateAmountOfPosts(authorId, amountChange);
     }
 }
